requirements.txt
streamlit
pandas 
plotly-express
watchdog
streamlit-dynamic-filters
streamlit-echarts 
nbformat
streamlit_card
langchain 
openai
python-environ 
tabulate
openpyxl
ezodf
lxml
odfpy